Revert "pippin ctx patches v2"
This reverts commit 00f02995eb42a3677426786c5f6fbd30418a6b12.

+++ components/ctx/ctx.c
@@ -13,7 +13,7 @@
 
 int mp_ctx_vfs_load_file (const char     *path,
                           unsigned char **contents,
-                          size_t         *length,
+                          long           *length,
                           long            max_length);
 #define CTX_LOAD_FILE mp_ctx_vfs_load_file
